requirements.txt
# Core data processing
numpy>=1.24.0
pandas>=2.0.0

# Geospatial processing
rasterio>=1.4.3
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0

# Data formats and API
requests>=2.31.0
pyyaml>=6.0.0

# Earth Observation specific
sentinelsat
boto3>=1.26.0  # For AWS S3 access

# Visualization
matplotlib>=3.7.0

# Development tools
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0

# FastAPI and related
fastapi==0.68.1
uvicorn==0.15.0
aiohttp==3.8.1
pydantic==1.8.2

# landsatxplore for the future
